requirements.txt
Django==1.6.5
Fabric==1.9.0
IPy==0.81
Markdown==2.4.1
Pillow==2.5.1
PyYAML==3.11
Pygments==1.6
South==1.0
Unidecode==0.04.16
argparse==1.2.1
beautifulsoup4==4.3.2
boto==2.32.1
certifi==14.05.14
chardet==2.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-annoying==0.8.0
django-appconf==0.6
django-bmemcached==0.2.1
django-bootstrap-pagination==1.5.0
django-braces==1.4.0
django-compressor==1.4
django-cors-headers==0.12
django-crispy-forms==1.4.0
django-extensions==1.3.10
django-filter==0.7
django-gravatar2==1.1.4
django-heroku-memcacheify==0.5
django-heroku-postgresify==0.3
django-herokuify==1.0.pre3
django-libsass==0.2
django-markdown==0.6.1
django-modelcluster==0.3
django-pylibmc-sasl==0.2.4
django-reporting==0.21
django-sslserver==0.12
django-storages==1.1.8
django-subdomains==2.0.4
django-taggit==0.12
django-templated-email==0.4.9
django-treebeard==2.0
django-user-agents==0.2.2
django-wysiwyg-redactor==0.3.9.1
djangorestframework==2.3.14
djrill==1.1.1
ecdsa==0.11
extras==0.0.3
gunicorn==19.1.0
html5lib==0.999
iso8601==0.1.10
libsass==0.5.0
newrelic==2.22.1.20
oauthlib==0.6.3
paramiko==1.14.0
pip-tools==0.3.4
psycopg2==2.5.3
pycrypto==2.6.1
pygments-json==0.1
raven==5.0.0
requests==2.3.0
six==1.7.3
sorl-thumbnail==11.12
static3==0.5.1
unicodecsv==0.9.4
wagtail==0.5
wsgiref==0.1.2
